% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctable_from_bams.R
\name{get_ctable_from_bams}
\alias{get_ctable_from_bams}
\title{Fetch co-occurrence count table (ctable) of GCH protection states from BAM files}
\usage{
get_ctable_from_bams(
  bamfiles,
  samplenames,
  regions,
  genome,
  alignerUsed = c("QuasR", "Bismark", "BISCUIT"),
  collapseBySample = TRUE,
  max_spacing = 300,
  remove_nonunique = TRUE,
  clip_until_nbg = 1L,
  noclip_protect_frac_above = 0.9,
  max_bisC_meth = 0.1,
  min_bisC_size = 10,
  mapqMin = 0,
  mapqMax = 255,
  absIsizeMin = NULL,
  absIsizeMax = NULL,
  min_read_size = 25L,
  max_read_size = 1000L,
  ncores = 1L
)
}
\arguments{
\item{bamfiles}{\code{character} vector with paths to bam files}

\item{samplenames}{\code{character} names of samples. If \code{collapseBySample} is \code{TRUE} counts are aggregated across bam files with same sample name.}

\item{regions}{\linkS4class{GRanges} object for regions of interest.}

\item{genome}{\code{character} path to fasta file which was used as reference for alignment.}

\item{alignerUsed}{\code{character} that defines which aligner was used to generate BAM files. Currently supported aligners are QuasR, Bismark and BISCUIT.}

\item{collapseBySample}{\code{logical} indicating whether to collapse counts for bam files with same \code{samplenames}. If \code{FALSE} prefix \code{s} followed by index is added to \code{samplenames}.}

\item{max_spacing}{maximum spacing length for which to collect co-occurrence counts.}

\item{remove_nonunique}{\code{logical} if \code{TRUE} only unique fragments are analyzed. Uniqueness is defined by fragments start, end and methylation states of all cytosines.}

\item{clip_until_nbg}{\code{integer} controlling clipping partial footprints at both ends of fragments. Namely, protected states are erased from each end until \code{clip_until_nbg} unprotected states are met.
Set \code{clip_until_nbg=0L} to skip clipping preprocessing.}

\item{noclip_protect_frac_above}{\code{numeric} controlling fragment clipping, i.e. fragments which have fraction of protected states higher than \code{max_protect_frac} escape clipping. This parameter allows to avoid removal of fully protected fragments during clipping process.}

\item{max_bisC_meth}{\code{numeric} controlling removal of fragments with failed bisulfite conversion, i.e. fragments with fraction of methylated non-GCH, and non-WCG cytosines higher then \code{max_bisC_meth} are removed from the analysis to get rid of fragments for which bisulfite conversion failed.}

\item{min_bisC_size}{\code{integer} setting minimum number non-GCH, and non-WCG cytosines for filtering according to bisulfite conversion efficiency, 
i.e. only fragments with number of non-GCH, and non-WCG cytosines higher then \code{min_bisC_size} are subject to filtering based on bisulfite conversion efficiency.}

\item{mapqMin}{\code{integer} setting minimum MAPQ of reads to be included into analysis.}

\item{mapqMax}{\code{integer} setting maximum MAPQ of reads to be included into analysis.}

\item{absIsizeMin}{\code{integer} or \code{NULL} (default). For paired-end experiments, minimal absolute insert
size (TLEN field in SAM Spec v1.4) of alignments to be included.}

\item{absIsizeMax}{\code{integer} or \code{NULL} (default). For paired-end experiments, maximal absolute insert
size (TLEN field in SAM Spec v1.4) of alignments to be included.}

\item{min_read_size}{\code{integer}. minimum read length to be included.}

\item{max_read_size}{\code{integer}. maximum read length to be included.}

\item{ncores}{number of cores to use.}
}
\value{
\code{tibble} where each row corresponds to sample - region combination. 
Columns of the returned \code{tibble} represent:
\describe{
  \item{SampleName}{Name of a sample as defined by \code{samplenames}}.
  \item{seqnames, start, end}{seqnames (or chromosomes) and coordinates of corresponding regions.}
  \item{names}{the same as \code{names(regions)} if it is not \code{NULL} or index of a corresponding region in \code{regions}.}
  \item{nFragsFetched}{number of fetched fragments before filtering.}
  \item{nFragsNonUnique}{number of non-unique fragments as defined by coordinates and methylation profiles.}
  \item{nFragsBisFailed}{number of fragments for which bisulfite conversion may have failed, namely non-GCH and non-WCG methylation level is higher than \code{max_bisC_meth}.}
  \item{nFragsAnalyzed}{number of fragments analysed for co-occurrence table after filtering non-unique and bisulfite conversion failed fragments.}
  \item{ClippedUntilNbg}{parameter \code{clip_until_nbg} used for clipping partial footprints from both ends.}
  \item{NotClippedProtectAbove}{parameter \code{noclip_protect_frac_above} used for clipping partial footprints from both ends.}
  \item{nFragsClipped}{number of fragments clipped from both ends to get rid of paetial footprints.}
  \item{CoocCountTable}{\code{matrix} of dimension \code{max_spacing}x4 for number of observed combination of protection states at different window sizes (distances). 
  Row names correspond to window size and columns contain number of observed 0-0, 0-1, 1-0 and 1-1. Note that first row contains counts for window of size 1 (or distance 0) and equal to total number of 0s and 1s in the data.}
}
}
\description{
Given a set of bamfiles and regions of interest this function searches for C to T and G to A mismatches in sequencing reads for several tri-nucleotide contexts
and collects statistics of observing 0-0, 0-1, 1-0 and 1-1 at different distances.
}
